// packages/database/src/internal/types.ts
export type UpsertOpenAIAccountInput = {
  userId?: string | null
  portalUserId?: string | null
  name?: string | null
  email: string
  picture?: string | null
  accountId?: string | null
  accountUserRole?: string | null
  workspaceName?: string | null
  planType?: string | null
  teamMemberCount?: number | null
  teamExpiresAt?: string | Date | null
  workspaceIsDeactivated?: boolean
  workspaceCancelledAt?: string | Date | null
  status?: string | null
  isShared?: boolean
  systemCreated?: boolean
  proxyId?: string | null
  accessToken?: string | null
  refreshToken?: string | null
  sessionToken?: string | null
  password?: string | null
  type?: string | null
  cooldownUntil?: string | Date | null
  rateLimit?: Record<string, unknown> | null
}

export type UpsertTeamAccountInput = {
  userId?: string | null
  portalUserId?: string | null
  ownerId?: string | null
  name?: string | null
  email: string
  accountId?: string | null
  accountUserRole?: string | null
  workspaceName?: string | null
  planType?: string | null
  teamMemberCount?: number | null
  teamExpiresAt?: string | Date | null
  workspaceIsDeactivated?: boolean
  workspaceCancelledAt?: string | Date | null
  status?: string | null
  systemCreated?: boolean
  proxyId?: string | null
  accessToken?: string | null
  refreshToken?: string | null
  password?: string | null
  type?: string | null
  cooldownUntil?: string | Date | null
  rateLimit?: Record<string, unknown> | null
}